add: 补充国际化及错码

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/atom/vm/DispatchVMStartupTaskAtom.kt
@@ -131,7 +131,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val pipelineInfo = pipelineRepositoryService.getPipelineInfo(projectId, pipelineId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_NOT_EXISTS.toInt(),
                 errorMsg = "流水线不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -141,7 +141,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val model = pipelineBuildDetailService.getBuildModel(buildId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_MODEL_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_MODEL_NOT_EXISTS.toInt(),
                 errorMsg = "流水线模型不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -151,7 +151,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
         val container = model.getContainer(vmSeqId)
             ?: throw BuildTaskException(
                 errorType = ErrorType.SYSTEM,
-                errorCode = ERROR_PIPELINE_NODEL_CONTAINER_NOT_EXISTS,
+                errorCode = ERROR_PIPELINE_NODEL_CONTAINER_NOT_EXISTS.toInt(),
                 errorMsg = "流水线的模型中指定构建容器${vmNames}不存在",
                 pipelineId = pipelineId,
                 buildId = buildId,
@@ -342,7 +342,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
     private fun agentException(pipelineId: String, buildId: String) =
         BuildTaskException(
             errorType = ErrorType.SYSTEM,
-            errorCode = ERROR_PIPELINE_AGENT_STATUS_EXCEPTION,
+            errorCode = ERROR_PIPELINE_AGENT_STATUS_EXCEPTION.toInt(),
             errorMsg = "第三方构建机的Agent状态异常",
             pipelineId = pipelineId,
             buildId = buildId,
